Add files via upload
These files contain the initial testing of operation of the EV3 gyro (used because it outputs angular position), and the GrovePi ultrasonic sensor (not functional on Pi 2 as of commit for some reason).  Additionally, they are modularized into their own .py files that can be called later in the main code.